requirements.txt
requests==2.31.0
websocket-client==1.7.0
colorama==0.4.6 